requirements.txt
Django>=3.0.6, <3.1.0
djangorestframework>=3.11.0, <3.12.0
flake8>=3.8.0, <3.9.0
psycopg2>=2.8.5, <2.9.0
djoser>=2.0.1, <2.1.0
drf-yasg>=1.17.1, <1.18.1